feat(exchange-core): product independent matching engine
BREAKING CHANGE: IREC related product and matching functions moved to @energyweb/exchange-core-irec package
