+++ packages/exchange-core/src/IMatchableProduct.ts
@@ -1,0 +1,5 @@
+export interface IMatchableProduct<TProduct, TProductFilter> {
+    product: TProduct;
+    matches(product: TProduct): boolean;
+    filter(filter: TProductFilter): boolean;
+}
