--- packages/exchange-core/src/Operator.ts
@@ -1,5 +1,0 @@
-export enum Operator {
-    EqualsTo,
-    GreaterThanOrEqualsTo,
-    LessThanOrEqualsTo
-}
