--- packages/exchange-core/src/TimeRange.ts
@@ -1,9 +1,0 @@
-export class TimeRange {
-    public readonly from: Date;
-
-    public readonly to: Date;
-
-    public static validate(timeRange: TimeRange) {
-        return timeRange.from < timeRange.to;
-    }
-}
