+++ packages/exchange-core/src/index.ts
@@ -1,13 +1,8 @@
+export * from './DirectBuy';
+export * from './IMatchableProduct';
 export * from './MatchingEngine';
 export * from './Order';
-export { Bid } from './Bid';
-export { Ask } from './Ask';
-export { Product } from './Product';
-export { ProductFilter, Filter } from './ProductFilter';
-export { DeviceVintage } from './DeviceVintage';
-export { Operator } from './Operator';
-export { DirectBuy } from './DirectBuy';
-export { TimeRange } from './TimeRange';
-export { Trade } from './Trade';
-export { TradeExecutedEvent } from './TradeExecutedEvent';
-export * from './MatchingEngineFactory';
+export * from './Testing';
+export * from './Trade';
+export * from './TradeExecutedEvent';
+export * from './strategy';
