+++ packages/exchange-core/src/strategy/AskPriceStrategy.ts
@@ -1,9 +1,9 @@
+import { IOrder } from '../Order';
 import { IPriceStrategy } from './IPriceStrategy';
-import { Ask, Bid } from '..';
 
 export class AskPriceStrategy implements IPriceStrategy {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    pickPrice(ask: Ask, bid: Bid): number {
+    pickPrice(ask: IOrder, bid: IOrder): number {
         return ask.price;
     }
 }
