+++ packages/exchange-core/src/strategy/IPriceStrategy.ts
@@ -1,5 +1,5 @@
-import { Ask, Bid } from '..';
+import { IOrder } from '../Order';
 
 export interface IPriceStrategy {
-    pickPrice(ask: Ask, bid: Bid): number;
+    pickPrice(ask: IOrder, bid: IOrder): number;
 }
